+++ src/library/URLComponent/URLPath.ts
@@ -6,7 +6,7 @@ import {
 
 import type {
   StringParsable,
-} from '@/library/Parser/string';
+} from '@/library/Parser/string/exports';
 
 import StringSubset from '@/library/StringSubset';
